requirements.txt
Django==5.2.1
djangorestframework==3.15.1
python-dotenv==1.0.1
django-cors-headers==4.3.1
requests==2.32.3
dj-database-url>=1.0.0
